% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{align}
\alias{align}
\title{Align lists}
\usage{
align(...)
}
\arguments{
\item{...}{Lists to be merged. Elements of each list should be named to avoid confusion.}
}
\description{
Creates a table with similarly named items on the same line to compare elements of multiple lists
}
\examples{
x <- list(a = 2, b = 13, c = 8)
y <- list(b = 4, c = 8, d = 12)
align(x, y)
merge.list(x, y, priority = "last")
}
