% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{backup}
\alias{backup}
\alias{revert}
\title{Backup function}
\usage{
backup(obj)

revert(obj)
}
\arguments{
\item{obj}{Object to save}
}
\description{
Creates a backup of an object. Useful when tweaking code.
}
\section{Functions}{
\itemize{
\item \code{revert}: Revert to saved version of object.
}}

\examples{
x <- sample(1:5, 50, replace = T)
backup(x)
x <- jitter(x)
plot(x, x)
revert(x)
}
\seealso{
\code{\link{revert}}
}
