% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{browse_db}
\alias{browse_db}
\title{Browse Database}
\usage{
browse_db(server, db = NULL)
}
\arguments{
\item{server}{Name/alias of server(s) containing the database}

\item{db}{Name of database}
}
\description{
Creates a \code{list} object containing column names for each table in a database.
}
