% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{choose.colors}
\alias{choose.colors}
\title{Color Choosing Function}
\usage{
choose.colors(locate = 1, plot.it = T)
}
\arguments{
\item{locate}{Number of colors to select from the palette}

\item{plot.it}{Logical. If \code{FALSE}, no palette is plotted and \code{colors}() is printed.}
}
\description{
Displays a palette of colors allowing a specified number of them to be selected.
}
\examples{
x <- c(10, 40, 25, 50, 20, 15)
cols <- choose.colors(length(x))
barplot(x, col = cols)
}
