% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{chunkplot}
\alias{chunkplot}
\title{Chunk Plot Function}
\usage{
chunkplot(chunk, num = 1)
}
\arguments{
\item{chunk}{Name of the chunk where the plot is generated.}

\item{num}{If more than one plot are created in a chunk, \code{num} designates which plot to return}
}
\description{
Returns a plot generated in a designated markdown chunk. Useful when plot needs to be outside of the chunk (e.g. next to a table).
}
