% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{complete.dt}
\alias{complete.dt}
\title{Complete \code{data.table} combinations}
\usage{
complete.dt(dt, ..., replaceFUN = NULL)
}
\arguments{
\item{dt}{\code{data.table} object to modify}

\item{...}{\code{list} giving the names of columns to complete. Needs at least two character vectors.}

\item{replaceFUN}{If supplied, this gives values other than NA to remaining columns.}
}
\description{
Adds rows to a \code{data.table} object for missing combinations of factor levels.
Useful when using \code{geom_bar} in \code{ggplot2}, so that bars are the same width.
}
\examples{
dt <- data.table(Gender = c("M", "F", "F"), Response = c("Y", "Y", "N"), Count = 1:3)
ggplot(dt, aes(Gender, Count, fill = Response)) + geom_bar(position = "dodge", stat = "identity")

dt2 <- complete.dt(dt, "Gender", "Response")
ggplot(dt2, aes(Gender, Count, fill = Response)) + geom_bar(position = "dodge", stat = "identity")
}
