% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{convert.tz}
\alias{convert.tz}
\title{Time Zone Converter}
\usage{
convert.tz(x, tz, oldtz = "UTC")
}
\arguments{
\item{x}{Vector of times to be converted. If a \code{character} vector is supplied it is converted to \code{POSIXct}}

\item{tz}{New time zone}

\item{oldtz}{Old time zone. Defaults to "UTC"}
}
\description{
Converts time objects of class \code{POSIXct} from one time zone to another.
}
\examples{
# Times of vernal equinox
x <- c("2004-03-20 06:49:00", "2005-03-20 12:33:00", "2006-03-20 18:26:00", "2007-03-21 00:07:00", "2008-03-20 05:48:00", 
       "2009-03-20 11:44:00", "2010-03-20 17:32:00", "2011-03-20 23:21:00", "2012-03-20 05:14:00", "2013-03-20 05:02:00") 
(y <- convert.tz(x, tz = "US/Mountain"))
data.table(UTC = x, US_Mountain = y)
}
\seealso{
\code{\link{POSIXct}}, \code{\link{OlsonNames}}
}
