% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{count.na}
\alias{count.na}
\title{Count NA values}
\usage{
count.na(x)
}
\arguments{
\item{x}{Vector containing NA values.}
}
\description{
Counts the number entries that are NA.
}
\examples{
x <- c(1, 2, NA, 4, NA, NA, 7, NA)
count.na(x)

sapply(airquality, count.na)
}
