% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{darken}
\alias{darken}
\title{Color Modifying Functions}
\usage{
darken(col, factor = 1/3)
}
\arguments{
\item{col}{Vector of colors to modify}

\item{factor}{Numerical value between 0 and 1 giving how much to modify. 0 is no modification.}
}
\description{
Returns a color lighter/darker than the input
}
\examples{
x <- c("red", "green", "blue", "yellow")
plot(seq_along(x), rep(1, length(x)), pch = 19, cex = 4, col = darken(x), ylim = c(1, 4))
points(seq_along(x), rep(2, length(x)), pch = 19, cex = 4, col = x)
points(seq_along(x), rep(3, length(x)), pch = 19, cex = 4, col = lighten(x, factor = 2/3))
points(seq_along(x), rep(4, length(x)), pch = 19, cex = 4, col = lighten(x, factor = 11/12))
}
