% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{dfSummary}
\alias{dfSummary}
\title{Data Table Summary}
\usage{
dfSummary(..., table.names = NULL, track = T)
}
\arguments{
\item{...}{\code{data.table} objects to be summarized.}

\item{table.names}{Vector of names to be associated with tables. If not supplied, the object names are used.}

\item{track}{Logical. Whether to display information while summarizing tables. Default is TRUE.}
}
\description{
Creates a table of summary statistics for \code{data.table} objects, each row coresponding to a column.
}
\examples{
dfSummary(data.table(iris))
}
