% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{dollarfull}
\alias{dollarfull}
\title{Dollar format}
\usage{
dollarfull(x)
}
\arguments{
\item{x}{Vector of numbers to format}
}
\description{
Uses code from \code{dollar} in \code{scales} package, but adds the option to not have even dollars rounded.
}
\examples{
x <- c(1, 1.5, 150)
dollarfull(x)
}
