% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{ext.chisq.test}
\alias{ext.chisq.test}
\title{Extended Chi-Square Test}
\usage{
ext.chisq.test(arr)
}
\arguments{
\item{arr}{An \code{array} or \code{matrix} object.}
}
\description{
Performss a categorical chi-squared test for tables of any size. Similar to \code{chisq.test}
}
\examples{
From \\code{\\link{lm}} documentation:
A <- cbind(1:4, (1:4)*2, (1:4)*3)
B <- cbind(1:4, (1:4)*2, c(3, 5, 2, 4))
}
\seealso{
\code{\link{chisq.test}}
}
