% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{fixedheaders}
\alias{fixedheaders}
\title{HTML Table with Fixed Header}
\usage{
fixedheaders()
}
\arguments{
\item{classname}{Default is 'fixedheader'. Class name of tables that should have fixed headers.}
}
\description{
Writes CSS and Javascript for a fixed table header for tables of a specified class.
}
\examples{
dt <- data.frame(x = 1:100, y = rnorm(100))
# Without fixed header
tab <- htmlTable(dt)
htmltools::tags$div(tab, style = "height: 300px; overflow-y: scroll;")
# With fixed header (only works when knit, otherwise the CSS is not used)
tab <- htmlTable(dt, css.class = "fixedheader")
htmltools::tags$div(tab, style = "height: 300px; overflow-y: scroll;")
}
