% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{geom_polyline}
\alias{geom_polyline}
\title{Plot Polynomial Curve}
\usage{
geom_polyline(mapping = NULL, data = NULL, precision = 100, ..., c0,
  na.rm = FALSE, show.legend = NA)
}
\description{
Creates a polynomial line, similar to \code{geom_abline}
}
\examples{
dat <- data.table(x = rnorm(1e3))[, y := rnorm(1e3, x, .2)^2][, x2 := x^2]
coefs <- lm(y ~ ., data = dat)$coef

ggplot(data.frame(x = x, y = rnorm(1e3, x, .2)^2), aes(x, y)) + geom_point() + geom_polyline(aes(c0 = coefs[1], c1 = coefs[2], c2 = coefs[3]))
}
