% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{geom_violin2}
\alias{geom_violin2}
\title{Modified Violin Plot}
\usage{
geom_violin2(mapping = NULL, data = NULL, stat = "ydensity",
  position = "identity", ..., draw_quantiles = NULL, trim = TRUE,
  scale = "area", na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\description{
Creates a violin plot with different distributions for each side. Use similar to \code{geom_violin}
with a two-level factor variable in the \code{group} or \code{fill} arguments.
}
\examples{
# geom_violin from ggplot2 package
ggplot(survey, aes(Fold, Wr.Hnd, fill = Sex)) + geom_violin()
# New geom_violin2 combined graphs
ggplot(survey, aes(Fold, Wr.Hnd, fill = Sex)) + geom_violin2()
}
