% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{gg_us_hex}
\alias{gg_us_hex}
\title{US Hex Map}
\usage{
gg_us_hex(data, mapping, dc = T, ...)
}
\arguments{
\item{dc}{Logical, whether or not to include a row for Washinton D.C.}
}
\description{
Generates a \code{data.table} object with center points and polygon information for graphing US states in a hex map.
}
\examples{
dt <- data.table(statepop)[, list(StateCd = abbr, pop_2015)]
gg_us_hex(dt, aes(fill = pop_2015))
}
