% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{gganimate}
\alias{gganimate}
\title{Animated plot}
\usage{
gganimate(..., n = 20)
}
\arguments{
\item{...}{\code{ggplot2} objects to transition between. At least 2 needed.}

\item{n}{Number of frames between plots. Default is 20.}
}
\description{
Returns a sequence of plots that transition between given plots.
}
\examples{

}
