% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{ggprop}
\alias{ggprop}
\title{Check Predicted Probabilities and Linearity of Means}
\usage{
ggprop(data, mapping, link = binomial(), predicted = T,
  stat = "identity", position = "identity", bins = 10, bound = T,
  ..., pt... = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{data}{\code{data.frame} as passed to \code{ggplot}()}

\item{mapping}{Aesthetic mapping using \code{aes}() as in \code{ggplot}. \code{x} and \code{y} are both required.}

\item{link}{Provides a link function for the binned means, if \code{predicted} is \code{FALSE}.}

\item{predicted}{Whether the given \code{x} represents a predicted probability. Default is \code{TRUE}.}

\item{stat, }{position, na.rm, show.legend, inherit.aes, ... Parameters that get passed through to \code{geom_errorbar}.}

\item{bins}{The number of quantiles to bin predicted probabilities into. Default is 10.}

\item{bound}{Whether to bound confidence intervals to the range [0, 1]. Default is \code{TRUE}}

\item{pt...}{\code{list} object with parameters to pass through to \code{geom_point}.}
}
\description{
Compares the proportion of success to a predicted probability using \code{ggplot2} functions. 
Alternatively, provides a check of linearity relative to a given covariate.
}
\examples{
esoph_mod <- glm(cbind(ncases, ncontrols) ~ ., data = esoph, family = "binomial")
esoph2 <- cbind(esoph, pred_prob = gtools::inv.logit(predict(esoph_mod)))
ggprop(esoph2, aes(pred_prob, cbind(ncases, ncontrols)))
ggprop(esoph2, aes(pred_prob, cbind(ncases, ncontrols)), predicted = F)
ggprop(esoph2, aes(pred_prob, cbind(ncases, ncontrols)), predicted = F, link = "identity")
}
