% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{ggroc}
\alias{ggroc}
\alias{ggroc2}
\title{\code{ggplot2} ROC Curves}
\usage{
ggroc(rc, labels = NULL, print.auc = c("title", "plot"),
  mark.sens = NULL, mark.spec = NULL, ...)

ggroc2(..., labels = NULL, print.auc = T, mark.sens = NULL,
  mark.spec = NULL)
}
\arguments{
\item{rc}{\code{\link[pROC]{roc}} object to evaluate.}

\item{labels}{}

\item{print.auc}{Where to print the area under the curve (AUC). By default it is put in the title.}

\item{mark.sens, }{mark.spec Sensitivites and specificities to highlight with plotted lines.}
}
\description{
Plots a ROC curve using \code{ggplot2}.
}
\section{Functions}{
\itemize{
\item \code{ggroc2}: Creates a ROC graph using \code{ggplot2}
}}

\examples{
data(aSAH)
rc <- pROC::roc(aSAH$outcome, aSAH$s100b, levels=c("Good", "Poor"))
ggroc(rc)
ggroc(rc, print.auc = "plot")
}
\seealso{
\code{\link{ggroc}}
}
