% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{ggsurv}
\alias{ggsurv}
\title{Kaplan-Meyer Plot for \code{ggplot2}}
\usage{
ggsurv(s, CI = "def", plot.cens = T, surv.col = "gg.def",
  cens.col = NULL, lty.est = 1, lty.ci = 2, cens.shape = 3,
  back.white = F, xlab = "Time", ylab = "Survival", main = "")
}
\arguments{
\item{s}{Survival model as returned from \code{survfit} in the \code{survival} package.}
}
\description{
Creates KM curves for a survival model using \code{ggplot2} functions.
}
