% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{grep.list}
\alias{grep.list}
\title{List Grep}
\usage{
grep.list(pattern, x, ...)
}
\arguments{
\item{pattern, }{... Parameters passed to \code{grep}.}

\item{x}{List of vectors to perform \code{grep} on.}
}
\description{
Performs \code{grep} on each entry in a list
}
\examples{
lst <- list(letters, LETTERS)
grep.list("a", lst)
grep.list("a", lst, value = T)
grep.list("a", lst, ignore.case = T)
grep.list("a", lst, value = T, ignore.case = T)
}
