% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{hardcode}
\alias{hardcode}
\title{Hard Code Objects}
\usage{
hardcode(x, assign = T, assign_oper = "<-", print = T)
}
\arguments{
\item{x}{Vector to hard code.}
}
\description{
Generates R code to re-create the given input as is. 
Generated code, when run, should create an identical object.
Supported object types include \code{numeric}, \code{character}, \code{logical}, \code{factor}, \code{matrix}, \code{data.frame} and \code{list}.
(other than \code{data.frame} objects which are converted to \code{data.table} objects).
}
\examples{
x <- rnorm(10, 5, 3)
hardcode(x)

lets <- letters
hardcode(y)

hardcode(mtcars)
}
