% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{html_chunk}
\alias{html_chunk}
\title{HTML Chunk}
\usage{
html_chunk(id, divs, labels)
}
\arguments{
\item{id}{Character, id attribute for html tags.}

\item{divs}{Character vector giving the ids of the div tags to be selected}

\item{labels}{Character vector of labels to display in dropdown menu}
}
\description{
Creates a section of html code that is modified according to selected inputs.
}
