% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{html_list}
\alias{html_list}
\title{HTML List Function}
\usage{
html_list(strings, details = NULL, detailpref = "- ", kind = c("ul", "ol",
  "dl"), type = c(1, "A", "a", "I", "i"))
}
\arguments{
\item{x}{Vector of values to list. Will be converted to \code{character}}
}
\description{
Creates HTML lists containing the elements of a vector.
}
\examples{
x <- paste("Item", 1:8)

# Unordered list (default)
html_list(x)

# Ordered list
html_list(x, kind = "dl")
}
