% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{html_model}
\alias{html_model}
\title{HTML Model Formula Function}
\usage{
html_model(mod, caption = NULL, ...)
}
\arguments{
\item{mod}{Model object.}
}
\description{
Creates a formula for a fitted model.
}
