% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{html_summary}
\alias{html_summary}
\title{HTML Model Summary Function}
\usage{
html_summary(mod, names = NULL, caption = NULL, ...)
}
\arguments{
\item{mod}{Model object.}

\item{names}{Optional 4-element vector of column names.}
}
\description{
Creates an \code{htmlTable} object for a model coefficient table.
}
