% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{ini}
\alias{ini}
\title{Open .odbc.ini file.}
\usage{
ini(print = F)
}
\arguments{
\item{print}{Boolean. Indicates whether the file should be displayed in the console. Default is FALSE.}
}
\description{
This function opens the .odbc.ini file which shows the servesr/databases that are available.
}
\examples{
ini()
}
