% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{is.color}
\alias{is.color}
\title{Color Checking/Formatting Functions}
\usage{
is.color(x)
}
\arguments{
\item{x}{Vector of \code{character} values to check.}
}
\description{
Checks whether character strings can be interpretted as colors. \code{is.color} is adapted from \code{\link{network:is.color}}
}
\examples{
cols <- c("red1", "red5", "#34AF0B", "#FFFFFF", "#FFFFFG", "#12345", "#123456", "#1234567", "#12345678")
x <- seq_along(cols)
plot(x, pch = 19, cex = 4, col = ifelse(is.color(cols), cols, "black"))
hex(cols)
data.table(String = cols, IsColor = is.color(cols), HexColor = hex(cols))
}
