% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{locateIndex}
\alias{locateIndex}
\title{Locate Plotted Point}
\usage{
locateIndex(dat, gg = F, print = T, n = 1, ...)
}
\arguments{
\item{dat}{\code{data.frame} with plotting coordinates given in columns \code{x} and \code{y}. If none of the columns have those names, the first two columns are used.}

\item{gg}{Whether to use \code{ggplot}. NOT CURRENTLY WORKING.}

\item{print}{Whether to print the row(s) corresponding to selected points.}

\item{n}{How many points to select.}

\item{...}{Parameters to be passed to \code{plot}.}
}
\description{
Uses \code{locator} to find the index of a plotted point. Useful for identifying outliers.
}
\examples{
locateIndex(trees, n = 3, cex = 2)
}
