% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{matpaste}
\alias{matpaste}
\title{Matrix Paste Function}
\usage{
matpaste(..., sep = " ", collapse = NULL, dim = NULL)
}
\arguments{
\item{..., }{sep, collapse Parameters passed to \code{paste}}

\item{dim}{Dimension to use if not implied by inputs.}
}
\description{
Same as \code{paste} but retains form when applied to a matrix.
}
\examples{
mat <- matrix(0, nrow = 3, ncol = 3)
paste("Row: ", row(mat), ", ", "Col: ", col(mat), sep = "")
matpaste("Row: ", row(mat), ", ", "Col: ", col(mat), sep = "")
}
