% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{nps}
\alias{nps}
\alias{npsind}
\alias{npsprop}
\alias{moe}
\title{Net Promoter Score (NPS) Functions}
\usage{
nps(x)

npsind(x, as.factor = F)

npsprop(x)

moe(x)
}
\arguments{
\item{x}{Vector of responses, on an 11-point Likert scale (integers 0 to 10).}

\item{as.factor}{Whether to return indicators as a factor variable. Default is \code{FALSE}}
}
\description{
Functions for calculating NPS from responses to the net promoter score survey question:
"How likely are you to recommend [company/product] to a friend or colleague?"
}
\section{Functions}{
\itemize{
\item \code{nps}: Converts 0-10 scale to -1, 0, 1 indicators representing detractors, passives and promoters, respectively.

\item \code{npsind}: Returns the margin of error for the NPS. This is a half-width for a confidence interval.

\item \code{npsprop}: Returns the proportion of detractors, passives and pormoters

\item \code{moe}: Calculates the Net Promoter Score.
}}

\examples{
responses <- c(10, 9, 10, 3, 7, 8, 5, 9, 8, 6)
npsind(responses)
}
