% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{padding}
\alias{padding}
\title{Padding Function}
\usage{
padding(n = 1)
}
\arguments{
\item{n}{Size of padding to add. Defaults to 1, representing padding of 1em on left and right.}
}
\description{
Used to add padding to left and right \code{htmlTable} cell objects
}
