% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{qu}
\alias{qu}
\title{Create Character Vector}
\usage{
qu(vec)
}
\arguments{
\item{vec}{Vector}
}
\description{
Creates a character vector from a vector of unquoted characters.
Ignores any value assigned to variables.
Does not work for character strings containing spaces.
}
\examples{
a <- 1
c(a, b)     # Gives error
# All the same
qu(c(a, b))
qu(c(a, "b"))
qu(c("a", "b"))
}
