% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{r2}
\alias{r2}
\alias{adjr2}
\title{R Squared and Adjusted R Squared Functions}
\usage{
r2(model)

adjr2(model)
}
\arguments{
\item{model}{A model as returned from \code{lm}}
}
\description{
Calculates the R^2 (or adjusted R^2) statistic of a model. Uses correlation of predicted values to actual values.
}
\section{Functions}{
\itemize{
\item \code{adjr2}: Calculates the adjusted R^2 statistic of a model
}}

\examples{
# From lm documentation:
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)

r2(lm.D9)
adjr2(lm.D9)
}
