% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{read.odbc}
\alias{read.odbc}
\title{Read Database}
\usage{
read.odbc(db, dbTable = NULL, dbQuery = NULL, ...)
}
\arguments{
\item{db}{Name of database.}

\item{dbTable}{Name of table in database.}

\item{dbQuery}{Character vector of length one containing a query to run against the database.}
}
\description{
Function for reading data from databases.
}
