% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{reglines}
\alias{reglines}
\title{Plot Regression Lines}
\usage{
reglines(mod)
}
\arguments{
\item{mod}{Linear regression model as returned by \code{lm()}.}
}
\description{
Plots response variable against all explanatory variables with estimated regression lines.
}
\examples{
library(data.table)
n <- 1e3

# Three independent explanatory variables
coef <- c(2.1, -0.8, 1.2)
dt <- data.table(x1 = rnorm(n), x2 = rnorm(n, , 2), x3 = rnorm(n, 4))
dt[, Mean := cbind(x1, x2, x3) \%*\% coef]
dt[, y := rnorm(.N, Mean, .5)]
mod <- lm(y ~ x1 + x2 + x3, data = dt)
reglines(mod)

# Three correlated explanatory variables
coef <- c(2.1, -0.8, 1.2)
dt <- data.table(x1 = rnorm(n))
dt[, x2 := rnorm(.N, -.8 + .7*x1)]
dt[, x3 := rnorm(.N, 1.2 + 1.1*x1 + .5*x2)]
dt[, Mean := cbind(x1, x2, x3) \%*\% coef]
dt[, y := rnorm(.N, Mean, .5)]
mod <- lm(y ~ x1 + x2 + x3, data = dt)
reglines(mod)

# Additional interaction term
coef <- c(2.1, -0.8, 1.2, -0.9)
dt <- data.table(x1 = rnorm(n))
dt[, x2 := rnorm(.N, -.8 + .7*x1)]
dt[, x3 := rnorm(.N, 1.2 + 1.1*x1 + .5*x2)]
dt[, Mean := cbind(x1, x2, x3, x1*x2) \%*\% coef]
dt[, y := rnorm(.N, Mean, .5)]
mod <- lm(y ~ x1 + x2 + x3 + x1*x2, data = dt)
reglines(mod)

# Additional categorical variable
coef <- c(2.1, -0.8, 1.2, -0.9, 2)
dt <- data.table(x1 = rnorm(n))
dt[, x2 := rnorm(.N, -.8 + .7*x1)]
dt[, Group := letters[rbinom(.N, 2, .5) + 1]]
dt[, x3 := rnorm(.N, 1.2 + 1.1*x1 + .5*x2)]
dt[, Mean := cbind(x1, x2, x3, x1*x2, factor(Group)) \%*\% coef]
dt[, y := rnorm(.N, Mean, .5)]
mod <- lm(y ~ x1 + x2 + Group + x3 + x1*x2, data = dt)
reglines(mod)

# Interaction with poorly fitted model
coef <- c(2.1, -0.8, 1.2, -0.9, 2, -3)
dt[, Mean := cbind(x1, x2, x3, x1*x2, factor(Group), as.numeric(factor(Group))*x3) \%*\% coef]
dt[, y := rnorm(.N, Mean, .5)]
mod <- lm(y ~ x1 + x2 + Group*x3 + x3 + x1*x2, data = dt)
reglines(mod)

# Interaction with better model
coef <- c(2.1, -0.8, 1.2, -0.9, 2, -3)
dt[, Mean := cbind(x1, x2, x3, x1*x2, factor(Group), as.numeric(factor(Group))*x3) \%*\% coef]
dt[, y := rnorm(.N, Mean, .5)]
mod <- lm(y ~ (x1 + x2 + x3)*Group + x1*x2, data = dt)
reglines(mod)
}
