% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{reglines2}
\alias{reglines2}
\title{Plot Regression Lines and Curves}
\usage{
reglines2(mod)
}
\arguments{
\item{mod}{A linear regression model}
}
\description{
Expand on previous function to allow more complicated functions -- polynomials, exponentials, etc.
}
\examples{
reglines2(mod)
}
