% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{replace.na}
\alias{replace.na}
\title{Replace NA values}
\usage{
replace.na(x, repl = NULL, ...)
}
\arguments{
\item{x}{Vector with NAs to be replaced}

\item{repl}{Value used to replace NAs, or a function to calculate a new value. If \code{repl} is not
given, default replacements depend on the data type for \code{x}: 0 (numeric), "" (character) or FALSE (logical)}

\item{...}{Additional parameters to be passed to \code{repl} if it is a function.}
}
\description{
Replaces NA values in the given input.
}
\examples{
x <- c(10, 20, NA, 40, NA, 60, NA, NA)
replace.na(x)
replace.na(x, repl = -1)
replace.na(x, repl = function(x) seq_along(x))
}
