% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{revf}
\alias{revf}
\title{Reverse factor levels}
\usage{
revf(vec)
}
\arguments{
\item{vec}{Factor vector}
}
\description{
Creates a copy of a factor variable where the order of the levels is reversed. This is useful when 
using coord_flip() in ggplot2, puts factors in order top to bottom instead of bottom to top.
Does NOT reverse the order of the vector!
}
\examples{
x <- factor(letters[1:5])
levels(x)
levels(revf(x))

dt <- data.table(grp = x[sample(1:5, size = 200, replace = T)])
ggplot(dt, aes(grp)) + geom_barplot() + coord_flip()
ggplot(dt, aes(revf(grp))) + geom_barplot() + coord_flip()
}
