% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{roc.dt}
\alias{roc.dt}
\title{ROC \code{data.table} Function}
\usage{
\method{roc}{dt}(rc, cost.fn = 1, cost.fp = 1, plot = T)
}
\arguments{
\item{rc}{\code{roc} object to evaluate.}

\item{cost.fn, }{cost.fp Costs associated with false negatives and false positives. Both default to 1.}

\item{plot}{If \code{TRUE} (default), the ROC curve is plotted using \code{ggroc}.}
}
\description{
Creates a \code{data.table} with sensitivity, specificity and threshold values associated with a \code{roc} object.
}
\examples{
data(aSAH)
rc <- roc(aSAH$outcome, aSAH$s100b, levels=c("Good", "Poor"))
dt <- roc.dt(rc)
dt
}
\seealso{
\code{\link{ggroc}}
}
