% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{roc_dt}
\alias{roc_dt}
\title{ROC \code{data.table} Function}
\usage{
roc_dt(rc, cost.fn = 1, cost.fp = 1, plot = T)
}
\arguments{
\item{rc}{\code{\link[pROC]{roc}} object to evaluate.}

\item{cost.fn, }{cost.fp Costs associated with false negatives and false positives. Both default to 1.}

\item{plot}{If \code{TRUE} (default), the ROC curve is plotted using \code{ggroc}.}
}
\description{
Creates a \code{data.table} with sensitivity, specificity and threshold values associated with a \code{\link[pROC]{roc}} object from package \code{\link[pROC]{pROC}}.
}
\examples{
data(aSAH)
rc <- pROC::roc(aSAH$outcome, aSAH$s100b, levels=c("Good", "Poor"))
dt <- roc_dt(rc)
dt
}
\seealso{
\code{\link{ggroc}}
}
