% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{roundNumeric}
\alias{roundNumeric}
\title{Round \code{data.table}}
\usage{
roundNumeric(dt, digits = 1)
}
\arguments{
\item{dt}{\code{data.frame} object to be rounded.}

\item{digits}{Number of places after the decimal.}
}
\description{
Creates a new \code{data.table} object with numeric fields rounded
}
