% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{rsavvy}
\alias{rsavvy}
\title{Add "Savvy" Functions}
\usage{
rsavvy()
}
\description{
Adds functions that are in the savvy package, which is not accessible outside of Riley1/Riley2
}
\examples{
rsavvy()
read.odbc
}
