% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{scale_x_dayyear}
\alias{scale_x_dayyear}
\title{Scale DayYear}
\usage{
scale_x_dayyear(p)
}
\description{
Adjusts the scale of an axis representing time to show days and years. Used with \code{ggplot2}.
}
