% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownFunctions.R
\name{scroll}
\alias{scroll}
\title{HTML Scrolling}
\usage{
scroll(tab, height = "300px", width = "100\%")
}
\arguments{
\item{tab}{\code{htmlTable} object or other HTML code to insert in scroll window.}

\item{height}{Height of the \code{<div>} container on the page. Default is "300px"}

\item{width}{Width of the \code{<div>} container on the page. Default is "100%"}
}
\description{
Wraps HTML code in a \code{<div>} tag with style set to limit the space on the page. All overflow content is accessed by scrolling.
}
\examples{
# Create table
tab <- htmlTable(mtcars)
scroll(tab)
}
