% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{setwd_curr}
\alias{setwd_curr}
\title{Set working directory to file location}
\usage{
setwd_curr()
}
\description{
Finds the location of the current code file and sets it as the working directory.
The function cannot be run from the console.
}
