% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{snippets}
\alias{snippets}
\title{Snippets Function}
\usage{
snippets(type = c("r", "markdown", "css", "c_cpp", "html", "java",
  "javascript", "python", "sql", "stan"), edit = F)
}
\arguments{
\item{type}{File type of snippets to be returned. Default is "r", which gives snippets for .R files.}

\item{edit}{Whether to open the snippets file to edit. Default is \code{FALSE}.}
}
\description{
Print available snippets. Snippets are little keywords that can be used to write code. Example: in an R script, type ts<SHIFT-TAB>
These can be customized in the r.snippets file, accessible via the following function.
}
\examples{
snippets()
}
