% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\docType{data}
\name{ss_theme}
\alias{ss_theme}
\alias{ssgray}
\title{Savvysherpa Colors}
\format{An object of class \code{character} of length 1.}
\usage{
ssgray
ssblue
ssyellow
ssgreen
ssred
sspurple
ssteal
sscols
g + ss_theme    # g is a returned \code{ggplot2} graph.
}
\description{
Colors and theme matching those used in Savvysherpa reports.
}
\keyword{datasets}
