% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{untable}
\alias{untable}
\title{Untable Function}
\usage{
untable(tab)
}
\arguments{
\item{tab}{Table as returned from \code{table}}
}
\description{
Undoes what \code{table} does.
}
\examples{
data(Titanic)
dim(Titanic)
dt1 <- as.data.table(Titanic)
dt2 <- untable(Titanic)
}
