% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{upd}
\alias{upd}
\title{Update function}
\usage{
upd(width = 260)
}
\arguments{
\item{width}{Sets the width, in characters, of the output in the console window.}
}
\description{
This function initializes lv as .Last.value and sets the width of any printed output.
}
\examples{
upd(200)
3 + 4
lv
getwd()
lv
}
