% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{vTable}
\alias{vTable}
\title{Venn Table Function}
\usage{
vTable(..., prop = FALSE, sums = FALSE)
}
\arguments{
\item{...}{Vectors to compare}

\item{prop}{Whether to show proportions instead of counts. Default is FALSE.}

\item{sums}{Whether to show row and columns sums using \code{Table}. Default is FALSE.}
}
\description{
Similar to a Venn diagram. Gives the counts of unique entries that are shared between given inputs.
}
\examples{
x <- c(2, 3, 3, 10, 12, 13)
y <- 1:6
vTable(x, y)
vTable(x, y, prop = TRUE)
vTable(x, y, sums = TRUE)
}
\seealso{
\code{\link{Table}}
}
