% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{yn}
\alias{yn}
\title{Yes/No Converter}
\usage{
yn(x, type = c("YN", "PN", "TF", "yn", "pn", "tf"), as.factor = T)
}
\arguments{
\item{x}{Vector of 0s and 1s.}
}
\description{
Converts 0/1 to Yes/No
}
\examples{
x <- rbinom(100, 1, .4)
table(x)
table(yn(x))
}
