% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{convert.tz}
\alias{convert.tz}
\title{Time Zone Converter}
\usage{
convert.tz(x, tz, oldtz = "UTC")
}
\arguments{
\item{x}{Vector of times to be converted. If a \code{character} vector is supplied it is converted to \code{POSIXct}}

\item{tz}{New time zone}

\item{oldtz}{Old time zone. Defaults to "UTC"}
}
\description{
Converts time objects of class \code{POSIXct} from one time zone to another.
}
\seealso{
\code{\link{POSIXct}}, \code{\link{OlsonNames}}
}
