% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphing.R
\name{is.color}
\alias{is.color}
\alias{hex}
\alias{darken}
\alias{lighten}
\title{Functions for checking, modifying and formatting colors}
\usage{
is.color(x)

hex(col, alpha = 1)

darken(col, factor = 1/3)

lighten(col, factor = 1/3)
}
\arguments{
\item{x}{Vector of \code{character} values to check.}

\item{col}{Vector of colors to modify}

\item{factor}{Numerical value between 0 and 1 giving how much to modify. 0 is no modification.}
}
\description{
Checks whether character strings can be interpretted as colors. \code{is.color} is adapted from \code{\link[network]{is.color}}
\code{lighten} and \code{darken} return a color lighter/darker than the input.
}
\section{Functions}{
\itemize{
\item \code{is.color}: Checks whether the given input can be interpreted as a color.

\item \code{hex}: Changes color to hex format.

\item \code{darken}: Darkens a color

\item \code{lighten}: Lightens a color
}}

\examples{
# Checking and formatting colors.
cols <- c("red1", "red5", "#34AF0B", "#FFFFFF", "#FFFFFG", "#12345", "#123456", "#1234567", "#12345678")
x <- seq_along(cols)
plot(x, pch = 19, cex = 4, col = ifelse(is.color(cols), cols, "black"))
hex(cols)
data.table(String = cols, IsColor = is.color(cols), HexColor = hex(cols))

# Making colors lighter or darker.
x <- c("red", "green", "blue", "yellow")
plot(seq_along(x), rep(1, length(x)), pch = 19, cex = 4, col = darken(x), ylim = c(1, 4))
points(seq_along(x), rep(2, length(x)), pch = 19, cex = 4, col = x)
points(seq_along(x), rep(3, length(x)), pch = 19, cex = 4, col = lighten(x, factor = 2/3))
points(seq_along(x), rep(4, length(x)), pch = 19, cex = 4, col = lighten(x, factor = 11/12))
}
